+++ tlRender/lib/tlTimelineVk/RenderPrivate.h
@@ -44,11 +44,17 @@ namespace tl
         // For drawing
         std::string vertexDummy();
         std::string fragmentDummy();
+
         std::string vertexSTs();
         std::string fragmentSTs();
+
+        std::string vertexUSD();
+        std::string fragmentUSD();
+        
         std::string vertexSource();
         std::string vertex2Source();
         std::string vertex2NoUVsSource();
+
         std::string meshFragmentSource();
         std::string colorMeshVertexSource();
         std::string colorMeshFragmentSource();