+++ tlRender/lib/tlTimelineVk/USDTextureSlots.h
@@ -1,0 +1,8 @@
+#pragma once
+
+enum USDTextureSlots
+{
+    USD_Diffuse = 0,
+    USD_Specular = 1,
+    USD_Roughness = 2,
+};